Improve appearance of BonusTargetOrbStreak and now it will require 2 slices to be achieved. Also, take out useless code.

+++ Entities/BonusTargetOrbStreak.js
@@ -6,6 +6,7 @@ define(['CirclePhysicsBody', 'SynchronizedTimers', 'Entities/Entity', 'Custom Ut
         this._hitBox = new CircularHitBoxWithAlgorithm(position, p_radius, new SliceAlgorithm(position, p_radius, gl, EffectsManager));
         
         this._handler = EffectsManager.requestLightningOrbWithStreakEffect(false, gl, 20, position, {});
+        this._numSlicesNeededToDestroy = 2;
     }
     
     //inherit from Entity
@@ -18,25 +19,30 @@ define(['CirclePhysicsBody', 'SynchronizedTimers', 'Entities/Entity', 'Custom Ut
     
     BonusTargetOrbStreak.prototype.setPosition = function(newPosition){
         Entity.Entity.prototype.setPosition.call(this, newPosition);
-        
         this._hitBox.setPosition(newPosition);
     }
     
     BonusTargetOrbStreak.prototype._setPositionWithInterpolation = function(newPosition){
         Entity.Entity.prototype._setPositionWithInterpolation.call(this, newPosition);
-        
         this._hitBox.setPosition(newPosition);
     }
     
-    BonusTargetOrbStreak.prototype.setAchievementPercentage = function(percent){
-        this._handler.increaseGlow(percent / 3.0);
+    BonusTargetOrbStreak.prototype.destroyAndReset = function(callback){
+        Entity.Entity.prototype.destroyAndReset.call(this, callback);
+        this._numSlicesNeededToDestroy = 2;
     }
     
     BonusTargetOrbStreak.prototype.runAchievementAlgorithmAndReturnStatus = function(mouseInputObj, callback){
         if(this._hitBox.processInput(mouseInputObj)){
-            EventSystem.publishEventImmediately("entity_destroyed", {entity: this, type: "bonus"});
-            this.destroyAndReset(callback);
-            return true;
+            if(this._numSlicesNeededToDestroy === 1){
+                EventSystem.publishEventImmediately("entity_destroyed", {entity: this, type: "bonus"});
+                this.destroyAndReset(callback);
+                return true;
+            }else{
+                this._numSlicesNeededToDestroy--;
+                this._hitBox.resetAlgorithm();
+                this._handler.setLightningState(false);
+            }
         }
         
         return false;
@@ -44,6 +50,7 @@ define(['CirclePhysicsBody', 'SynchronizedTimers', 'Entities/Entity', 'Custom Ut
     
     BonusTargetOrbStreak.prototype.spawn = function(callback){
         Entity.Entity.prototype.spawn.call(this, callback);
+        this._handler.setLightningState(true);
         EventSystem.publishEventImmediately("entity_spawned", {entity: this, type: "bonus"});
     }
